Move to scriptData
