+++ lib/app.js
@@ -4,8 +4,7 @@ const request = require('./utils/request');
 const queryString = require('querystring');
 const cheerio = require('cheerio');
 const R = require('ramda');
-const matchScriptData = require('./utils/matchScriptData');
-const generateExtractFieldsFn = require('./utils/generateExtractFieldsFn');
+const scriptData = require('./utils/scriptData');
 
 const PLAYSTORE_URL = 'https://play.google.com/store/apps/details';
 
@@ -31,9 +30,9 @@ function app (opts) {
     }, opts.requestOptions);
 
     request(options, opts.throttle)
-      .then(matchScriptData)
+      .then(scriptData.parse)
     // comment next line to get raw data
-      .then(generateExtractFieldsFn(MAPPINGS))
+      .then(scriptData.extractor(MAPPINGS))
       .then(R.assoc('appId', opts.appId))
       .then(R.assoc('url', reqUrl))
       .then(resolve)
