+++ lib/similar.js
@@ -4,8 +4,7 @@ const request = require('./utils/request');
 const R = require('ramda');
 const url = require('url');
 const queryString = require('querystring');
-const matchScriptData = require('./utils/matchScriptData');
-const generateExtractFieldsFn = require('./utils/generateExtractFieldsFn');
+const scriptData = require('./utils/scriptData');
 
 const PLAYSTORE_URL = 'https://play.google.com/store/apps/details';
 
@@ -32,8 +31,8 @@ function similar (opts) {
     }, opts.requestOptions);
 
     request(options, opts.throttle)
-      .then(matchScriptData)
-      .then(generateExtractFieldsFn(PATH_MAPPING))
+      .then(scriptData.parse)
+      .then(scriptData.extractor(PATH_MAPPING))
       .then(data => {
         if (data.path === undefined) {
           throw Error('not found similar link');
@@ -43,10 +42,10 @@ function similar (opts) {
           followAllRedirects: true
         }, opts.requestOptions));
       })
-      .then(matchScriptData)
+      .then(scriptData.parse)
       .then(data => {
         const input = R.path(['ds:3', 0, 1, 0, 0, 0], data);
-        return R.map(generateExtractFieldsFn(MAPPINGS), input);
+        return R.map(scriptData.extractor(MAPPINGS), input);
       })
       .then(resolve)
       .catch(reject);
