+++ lib/utils/scriptData.js
@@ -1,11 +1,31 @@
 'use strict';
+
 const R = require('ramda');
+
+function extractor (mappings) {
+  /*
+  * Map the MAPPINGS object, applying each field spec to the parsed data.
+  * If the mapping value is an array, use it as the path to the extract the
+  * field's value. If it's an object, extract the value in object.path and pass
+  * it to the function in object.fun
+  */
+  return function extractFields (parsedData) {
+    return R.map((spec) => {
+      if (R.is(Array, spec)) {
+        return R.path(spec, parsedData);
+      }
+      // assume spec object
+      const input = R.path(spec.path, parsedData);
+      return spec.fun(input);
+    }, mappings);
+  };
+}
 
 /*
  * Extract the javascript objects returned by the AF_initDataCallback functions
  * in the script tags of the app detail HTML.
  */
-function matchScriptData (response) {
+function parse (response) {
   const scriptRegex = />AF_initDataCallback[\s\S]*?<\/script/g;
   const keyRegex = /(ds:.*?)'/;
   const valueRegex = /return ([\s\S]*?)}}\);<\//;
@@ -29,4 +49,4 @@ function matchScriptData (response) {
   }, {});
 }
 
-module.exports = matchScriptData;
+module.exports = Object.assign({parse, extractor});
